Implement starmash events interface minus theme/setting support
SWAM has a pretty simple event handling interface and, as I have access
to minified source code for starmash it is pretty straightforward to set
up something that exactly replicates the documented API.

I have tested this with Detect's shield timer and it was able to run
fine with the exception of it expecting the SWAM AutoPilot to be
available.

Note that there are a bunch of things that are _not_ implemented here:
- Themes. SWAM.registerExtension accepts a themes array but otherwise
  completely ignores it.
- Settings. I've added stub SettingsProvider and SettingsSection classes
  but they do not actually do anything.

Future work would be add support for settings and themes.